Small fixes in pump routine

+++ example/calcPhiASE.m
@@ -1,237 +1,216 @@
-% calcPhiASE
-% calculates the phi_ASE values for a given input
-%first, enter all parameters for the mesh:
-% p
-% normals_x
-% normals_y
-% forbidden
-% normals_p
-% sorted_int
-% t_int
-% z_mesh 
-% mesh_z
-% N_tot
-% beta_vol
-% laser                      struct: parameters for the laser (wavelength sigmas)
-% crystal                    struct: contains the crystal parameters (flourescence)
-% beta_cell
-% surface
-% x_center
-% y_center
-% NumRays                     the minimal number of rays to start for each samplepoint
-% clad_int
-% clad_number
-% clad_abs
-% refractiveIndices         the refractive indices for each surface [top_inner, top_outer, bottom_inner, bottom_outer]
-% reflectivities            the amount of reflection in case there is no total internal reflection)
-% MaxRays                   the maximum number of rays for each samplepoint
-% mse_threshold             the maximum allowed expectation (see calc_dndt_ase.cu) for each samplepoint (calculation will be restarted with more rays, if expectation not low enough)
-% bool use_reflections      if reflections should be used
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% calcPhiASE %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %
-% phi_ASE                   phi_ASE values, multiple Wavelengths in different columns
-% mse_values           real expectation-values for each samplepoint (aligned like phi_ASE)
-% N_rays                    number of rays that were used for each samplepoint
+% calculates the phiASE values for a given input
+% most meshing paramers are given through the function parameters.
+% However, many parameters for optimization of the computation are
+% set in the beginning of the function (adjust as needed)
+% 
+% for most mesh parameters see README file
+%
+% @return phiASE the ASE-Flux for all the given sample points
+% @return mseValues the MeanSquaredError values corresponding to phiASE
+% @return raysUsedPerSample the number of rays used to calculate each phiASE value
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+function [phiASE, mseValues, raysUsedPerSample] = calcPhiASE(
+  points,
+  trianglePointIndices,
+  betaCells,
+  betaVolume,
+  claddingCellTypes,
+  claddingNumber,
+  claddingAbsorption,
+  useReflections,
+  refractiveIndices,
+  reflectivities,
+  triangleNormalsX,
+  triangleNormalsY,
+  triangleNeighbors,
+  triangleSurfaces,
+  triangleCenterX,
+  triangleCenterY,
+  triangleNormalPoint,
+  forbiddenEdge,
+  minRaysPerSample,
+  maxRaysPerSample,
+  mseThreshold,
+  repetitions,
+  nTot,
+  thickness,
+  laserParameter,
+  crystal,
+  numberOfLevels,
+  runmode,
+  maxGPUs,
+  nPerNode
+  )
 
-%function [phi_ASE, mse_values, N_rays] = calcPhiASE(p,normals_x,normals_y,forbidden,normals_p,sorted_int,t_int,z_mesh,mesh_z,N_tot,beta_vol,laser,crystal,beta_cell,surface,x_center,y_center,NumRays,clad_int, clad_number, clad_abs, refractiveIndices, reflectivities,MaxRays,mse_threshold,use_reflections)
 
-function [phi_ASE, mse_values, N_rays] = calcPhiASE(p,t_int,beta_cell,beta_vol,normals_x,normals_y,sorted_int,surface,x_center,y_center,normals_p,forbidden,NumRays,N_tot,z_mesh,laser,crystal,mesh_z)
-
-%%% Added to the interface %%%
-%laser
-%crystal
-%mesh_z
-
-
-%%% Defined here in the file %%%
-%MaxRays
-%mse_threshold
-%use_reflections
-
-
-%%% are not really defined and will be created with dummy values %%% 
-%clad_int
-%clad_number
-%clad_abs
-%refractiveIndices
-%reflectivities
-
-MaxRays = 100000;
-mse=0.05;
-use_reflections = false; 
-MAX_GPUS=1;
-[a,b] = size(p);
+%%%%%%%%%%%%% auto-generating some more input %%%%%%%%%%%%%
 minSample=0;
-maxSample=(mesh_z*a)-1;
+[nP,b] = size(points);
+maxSample=(numberOfLevels*nP)-1;
 
-used_dummy = false;
-
-%create dummy variables
-if(~exist('mse_threshold','var'))
-  %WARNING = 'The variable mse_threshold does not exist'
-  [a,b] = size(laser.s_ems);
-  mse_threshold = ones(1,a)*mse;
-  used_dummy=true;
-end
-
-%create dummy variables
-if(~exist('clad_int','var') || ~exist('clad_num','var') ||  ~exist('clad_abs','var'))
-  %WARNING = 'The variables "clad_int", "clad_num" or "clad_abs" (or a combination) do not exist'
-  [a,b] = size(sorted_int);
-  clad_int = ones(1,a);
-  clad_number = 3;
-  clad_abs = 5.5;
-  used_dummy=true;
-end
-
-%create dummy variables
-if(~exist('reflectivities','var') || ~exist('refractiveIndices','var'))
-  %WARNING = 'The variables "reflectivities" or "refractiveIndices" (or both) do not exist'
-  refractiveIndices = [1.83,1,1.83,1];
-  [a,b] = size(sorted_int);
-  reflectivities = ones(1,a*2) * 0;
-  used_dummy=true;
-end
-
-% create the correct reflection-parameter for the c-function
-REFLECT='';
-if(use_reflections == true)
+if(useReflections == true)
     REFLECT=' --reflection';
+else
+    REFLECT='';
 end
 
-if(used_dummy == true)
-  disp([ 'WARNING: Some variables were set as dummies' ]);
+if(strcmpi(runmode,'mpi'))
+  Prefix=[ 'mpiexec -npernode ' num2str(nPerNode) ' ' ];
+  maxGPUs=1;
+else
+  Prefix='';
 end
 
-
-CURRENT_DIR = pwd;
+% create a tmp-folder in the same directory as this script
 FILENAME=[ mfilename('fullpath') '.m' ];
 [ CALCPHIASE_DIR, NAME , EXTENSION ] = fileparts(FILENAME);
+TMP_FOLDER=[ CALCPHIASE_DIR filesep 'input_tmp' ];
 
-% create all the textfiles in a separate folder
-TMP_FOLDER = [ '/' 'tmp' filesep 'calcPhiASE_tmp' ];
 
-% make sure that the input is clean 
+%%%%%%%%%%%%%%%%%% doing the computation %%%%%%%%%%%%%%%%%%
+% make sure that the temporary folder is clean 
 clean_IO_files(TMP_FOLDER);
 
-% create the new input based on the MATLAB variables
-create_calcPhiASE_input(p,normals_x,normals_y,forbidden,normals_p,sorted_int,t_int,z_mesh,mesh_z,N_tot,beta_vol,laser,crystal,beta_cell,surface,x_center,y_center,clad_int,clad_number,clad_abs,refractiveIndices,reflectivities,mse_threshold,TMP_FOLDER,CURRENT_DIR);
+% create new input in the temporary folder
+create_calcPhiASE_input(points,triangleNormalsX,triangleNormalsY,forbiddenEdge,triangleNormalPoint,triangleNeighbors,trianglePointIndices,thickness,numberOfLevels,nTot,betaVolume,laserParameter,crystal,betaCells,triangleSurfaces,triangleCenterX,triangleCenterY,claddingCellTypes,claddingNumber,claddingAbsorption,refractiveIndices,reflectivities,TMP_FOLDER);
 
+% do the propagation
+status = system([ Prefix CALCPHIASE_DIR '/bin/calcPhiASE ' '--mode=' runmode ' --rays=' num2str(minRaysPerSample) ' --maxrays=' num2str(maxRaysPerSample) REFLECT ' --input=' TMP_FOLDER ' --output=' TMP_FOLDER ' --min_sample_i=' num2str(minSample) ' --max_sample_i=' num2str(maxSample) ' --maxgpus=' num2str(maxGPUs) ' --repetitions=' num2str(repetitions) ' --mse-threshold=' num2str(mseThreshold) ]);
 
-  % do the propagation
-  system([ CALCPHIASE_DIR '/bin/calcPhiASE ' '--mode=ray_propagation_gpu ' '--rays=' num2str(NumRays) ' --maxrays=' num2str(MaxRays) REFLECT ' --experiment=' TMP_FOLDER ' --min_sample_i=' num2str(minSample) ' --max_sample_i=' num2str(maxSample) ' --maxgpus=' num2str(MAX_GPUS) ]);
-
-  % get the result
-  [ mse_values, N_rays, phi_ASE ] = parse_calcPhiASE_output(TMP_FOLDER,CURRENT_DIR);
-
-  % cleanup
-  %clean_IO_files(TMP_FOLDER);
-  % 
+if(status ~= 0)
+    error(['this step of the raytracing computation did NOT finish successfully. Aborting.']);
 end
 
-%takes all the variables and puts them into textfiles, so the CUDA code can parse them
-function create_calcPhiASE_input (p,normals_x,normals_y,forbidden,normals_p,sorted_int,t_int,z_mesh,mesh_z,N_tot,beta_vol,laser,crystal,beta_cell,surface,x_center,y_center,clad_int,clad_number,clad_abs,refractiveIndices,reflectivities,mse_threshold,FOLDER,CURRENT_DIR)
+% get the result
+[ mseValues, raysUsedPerSample, phiASE ] = parse_calcPhiASE_output(TMP_FOLDER);
+
+% cleanup
+clean_IO_files(TMP_FOLDER);
+
+end
+
+
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%% parse_calcPhiASE_output %%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+% takes all the variables and puts them into textfiles
+% so the CUDA code can parse them. Take care that the
+% names of the textfiles match those in the parsing function
+% 
+% for most parameters, see calcPhiASE (above)
+% @param FOLDER the folder in which to create the input files (usually a
+%               temporary folder visible by all the participating nodes)
+%
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+function create_calcPhiASE_input (points,triangleNormalsX,triangleNormalsY,forbiddenEdge,triangleNormalPoint,triangleNeighbors,trianglePointIndices,thickness,numberOfLevels,nTot,betaVolume,laserParameter,crystal,betaCells,triangleSurfaces,triangleCenterX,triangleCenterY,claddingCellTypes,claddingNumber,claddingAbsorption,refractiveIndices,reflectivities,FOLDER)
+CURRENT_DIR = pwd;
 
 mkdir(FOLDER);
 cd(FOLDER);
 
-x=fopen('p_in.txt','w');
-fprintf(x,'%.50f\n',p);
+x=fopen('points.txt','w');
+fprintf(x,'%.50f\n',points);
 fclose(x);
 
-x=fopen('n_x.txt','w');
-fprintf(x,'%.50f\n',normals_x);
+x=fopen('triangleNormalsX.txt','w');
+fprintf(x,'%.50f\n',triangleNormalsX);
 fclose(x);
 
-x=fopen('n_y.txt','w');
-fprintf(x,'%.50f\n',normals_y);
+x=fopen('triangleNormalsY.txt','w');
+fprintf(x,'%.50f\n',triangleNormalsY);
 fclose(x);
 
-x=fopen('forbidden.txt','w');
-fprintf(x,'%d\n',forbidden);
+x=fopen('forbiddenEdge.txt','w');
+fprintf(x,'%d\n',forbiddenEdge);
 fclose(x);
 
-x=fopen('n_p.txt','w');
-fprintf(x,'%d\n',normals_p);
+x=fopen('triangleNormalPoint.txt','w');
+fprintf(x,'%d\n',triangleNormalPoint);
 fclose(x);
 
-x=fopen('neighbors.txt','w');
-fprintf(x,'%d\n',sorted_int);
+x=fopen('triangleNeighbors.txt','w');
+fprintf(x,'%d\n',triangleNeighbors);
 fclose(x);
 
-x=fopen('t_in.txt','w');
-fprintf(x,'%d\n',t_int);
+x=fopen('trianglePointIndices.txt','w');
+fprintf(x,'%d\n',trianglePointIndices);
 fclose(x);
 
 % thickness of one slice!
-x=fopen('z_mesh.txt','w');
-fprintf(x,'%.50f\n',z_mesh);
+x=fopen('thickness.txt','w');
+fprintf(x,'%.50f\n',thickness);
 fclose(x);
 
 % number of slices
-x=fopen('mesh_z.txt','w');
-fprintf(x,'%d\n',mesh_z);
+x=fopen('numberOfLevels.txt','w');
+fprintf(x,'%d\n',numberOfLevels);
 fclose(x);
 
-x=fopen('size_t.txt','w');
-[a,b] = size(t_int);
+x=fopen('numberOfTriangles.txt','w');
+[a,b] = size(trianglePointIndices);
 fprintf(x,'%d\n',a);
 fclose(x);
 
-x=fopen('size_p.txt','w');
-[a,b] = size(p);
+x=fopen('numberOfPoints','w');
+[a,b] = size(points);
 fprintf(x,'%d\n',a);
 fclose(x);
 
-x=fopen('n_tot.txt','w');
-fprintf(x,'%.50f\n',N_tot);
+x=fopen('nTot.txt','w');
+fprintf(x,'%.50f\n',nTot);
 fclose(x);
 
-x=fopen('beta_v.txt','w');
-fprintf(x,'%.50f\n',beta_vol);
+x=fopen('betaVolume.txt','w');
+fprintf(x,'%.50f\n',betaVolume);
 fclose(x);
 
-x=fopen('sigma_a.txt','w');
-fprintf(x,'%.50f\n',laser.s_abs);
+x=fopen('sigmaA.txt','w');
+fprintf(x,'%.50f\n',laserParameter.s_abs);
 fclose(x);
 
-x=fopen('sigma_e.txt','w');
-fprintf(x,'%.50f\n',laser.s_ems);
+x=fopen('sigmaE.txt','w');
+fprintf(x,'%.50f\n',laserParameter.s_ems);
 fclose(x);
 
-x=fopen('tfluo.txt','w');
+x=fopen('crystalTFluo.txt','w');
 fprintf(x,'%.50f\n',crystal.tfluo);
 fclose(x);
 
-x=fopen('beta_cell.txt','w');
-fprintf(x,'%.50f\n',beta_cell);
+x=fopen('betaCells.txt','w');
+fprintf(x,'%.50f\n',betaCells);
 fclose(x);
 
-x=fopen('surface.txt','w');
-fprintf(x,'%.50f\n',surface);
+x=fopen('triangleSurfaces.txt','w');
+fprintf(x,'%.50f\n',triangleSurfaces);
 fclose(x);
 
-x=fopen('x_center.txt','w');
-fprintf(x,'%.50f\n',x_center);
+x=fopen('triangleCenterX.txt','w');
+fprintf(x,'%.50f\n',triangleCenterX);
 fclose(x);
 
-x=fopen('y_center.txt','w');
-fprintf(x,'%.50f\n',y_center);
+x=fopen('triangleCenterY.txt','w');
+fprintf(x,'%.50f\n',triangleCenterY);
 fclose(x);
 
-
-x=fopen('clad_int.txt','w');
-fprintf(x,'%d\n',clad_int);
+x=fopen('claddingCellTypes.txt','w');
+fprintf(x,'%d\n',claddingCellTypes);
 fclose(x);
 
-x=fopen('clad_num.txt','w');
-fprintf(x,'%d\n',clad_number);
+x=fopen('claddingNumber.txt','w');
+fprintf(x,'%d\n',claddingNumber);
 fclose(x);
 
-x=fopen('clad_abs.txt','w');
-fprintf(x,'%.50f\n',clad_abs);
+x=fopen('claddingAbsorption.txt','w');
+fprintf(x,'%.50f\n',claddingAbsorption);
 fclose(x);
 
-x=fopen('refractive_indices.txt','w');
+x=fopen('refractiveIndices.txt','w');
 fprintf(x,'%3.5f\n',refractiveIndices);
 fclose(x);
 
@@ -239,21 +218,29 @@ x=fopen('reflectivities.txt','w');
 fprintf(x,'%.50f\n',reflectivities);
 fclose(x);
 
-x=fopen('mse_threshold.txt','w');
-fprintf(x,'%.15f\n',mse_threshold);
-fclose(x);
-
 cd(CURRENT_DIR);
 end 
 
 
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%% parse_calcPhiASE_output %%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
 % takes the output from the CUDA code and fills it into a variable
-function [mseValues,  N_rays, phi_ASE] = parse_calcPhiASE_output (FOLDER,CURRENT_DIR)
+% assumes that the matrix is saved as a 3D-matrix where the first line 
+% denotes the dimensions and the second line is the whole data
+%
+% @param FOLDER the folder which contains the output files
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+function [mseValues,  raysUsedPerSample, phiASE] = parse_calcPhiASE_output (FOLDER)
+CURRENT_DIR = pwd;
 cd (FOLDER);
 fid = fopen('phi_ASE.txt');
 arraySize = str2num(fgetl(fid));
-phi_ASE = str2num(fgetl(fid));
-phi_ASE = reshape(phi_ASE,arraySize);
+phiASE = str2num(fgetl(fid));
+phiASE = reshape(phiASE,arraySize);
 fclose(fid);
 
 fid=fopen('mse_values.txt');
@@ -265,26 +252,38 @@ fclose(fid);
 
 fid = fopen('N_rays.txt');
 arraySize = str2num(fgetl(fid));
-N_rays = str2num(fgetl(fid));
-N_rays = reshape(N_rays,arraySize);
+raysUsedPerSample = str2num(fgetl(fid));
+raysUsedPerSample = reshape(raysUsedPerSample,arraySize);
 fclose(fid);
 
 cd (CURRENT_DIR);
 end
 
+
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%% clean_IO_files %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
 % deletes the temporary folder and the dndt_ASE.txt
+% 
+% @param TMP_FOLDER the folder to remove
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 function clean_IO_files (TMP_FOLDER)
-
-% disable warnings for this (if file is nonexistant...)
-s = warning;
-warning off all;
-
-A = exist(TMP_FOLDER,'dir');
-
+A = exist(TMP_FOLDER,'dir'); % continue only if the folder exists!
 if A == 7
+  s = warning;
+  warning off all;
+
+  isOctave = exist('OCTAVE_VERSION') ~= 0;
+  if(isOctave)
+    confirm_recursive_rmdir (0);
+  else
     rmdir(TMP_FOLDER,'s');
   end
 
   warning(s);
 end
 
+end
